+++ python/tests/test_init.py
@@ -1,5 +1,0 @@
-import bedrock_bio
-
-
-def test_version():
-    assert bedrock_bio.__version__