+++ r/R/zzz.R
@@ -1,9 +1,8 @@
 pkg <- new.env(parent = emptyenv())
 
 .onLoad <- function(libname, pkgname) {
-  pkg$domain <- Sys.getenv("BB_R2_DOMAIN", "data.bedrock.bio")
-  pkg$catalog_url <- paste0("https://", pkg$domain, "/catalog.json")
-  pkg$credentials_url <- paste0("https://", pkg$domain, "/credentials.json")
+  pkg$catalog_url <- "https://data.bedrock.bio/catalog.json"
+  pkg$credentials_url <- "https://data.bedrock.bio/credentials.json"
 }
 
 .onUnload <- function(libpath) {